// types/tarjetaOferta.ts
export type TarjetaOferta = {
    oferta_id: number; //Tabla Ofertas
    convenio_id: number; //Tabla Convenios
    nombre_emp: string; //Tabla Empresas
    logo_url: string; //Tabla Empresas
    direccion: string; //Tabla Sucursales
    ciudad: string; //Tabla Sucursales
    categoria_id: number; //Tabla Categorias
    nombre_cat: string; //Tabla Categorias
    titulo: string; //Tabla Convenios
    condiciones: string; //Tabla Convenios
}